Updates to Expedition + Cats
Implemented another database scriptableobject, CatStats, in order to assign the 3 stats for each cat type. Expeditions now account for these stats and check the success chance of each roll on completion, for now doing a debug log to show the reward tiers accordingly. It is an RNG system at the moment.